feat: allow state in HookSet
This is quite useful for example when doing statement coverage. One
could store a MPSC tx end inside a BranchHookSet in order to send the
current program counter on every branching instruction. Stateless are
still possible and should be zero cost, they can be implemented by just
using an empty struct on which the HookSet trait is implemented.