chore: make make_span fallible
This makes the `make_span` method fallible. It also removes the
TODO from the `impl Index for Span`, because Index can not easily be
made to return a `Result`, as it is defined to return a reference to
`Self::Output`. Thus, if one could only return a shared ref to a Result,
which probably has suboptimal ergonomics.

Signed-off-by: wucke13 <wucke13@gmail.com>