feat: add `RwSpinLock` implementation
In preparation for support of the multiple linear memories proposal, we
will need locking. This commit introduces a first version of a naive
read-write lock implementation that solely relies on spinning, avoiding
any dependence on an operating system. The lock is tailored for the use
case in a linear memory which is shared between multiple, concurrently
running interpreter instances.

Signed-off-by: wucke13 <wucke13+github@gmail.com>