fix: use panic handler in specification test only when required
Previously, the panic handler was set to do nothing at the start of the
test so that panics in the interpreter do not emit backtraces to STDOUT.
This prevents us from having actual useful panic backtraces, which result
from bugs in the specification test runner itself.
This commit reduces the time where no panic handler as much as possible
by removing it temporarily only when calls to the interpreter are made.

Signed-off-by: Florian Hartung <florian.hartung@dlr.de>