fix: msrv problem
Before this commit, the Rust version in the devShell would equal to the
configured MSRV, but the Rust version used for CI and the nix pkg would
be the most up-to-date Rust toolchain in the oxalica rust-overlay.

So CI would pass if MSRV is violated, but in the devShell `cargo build`
would fail.

---

This commit changes things quite a bit:

- A new check is added to the flake checks, which builds the interpreter
  with tests using the MSRV.
- The normal build of the wasm-interpreter however uses the Rust version
  found in the nixpkgs. The same version is used for the devShell.
- We do neither measure coverage nor run the benchmarks for the MSRV.
- The stale naersk input is removed from the `flake.nix`.
- `nix flake check` was set as a commit hook already. However, with the
  MSRV check package now there is an actual build of the interpreter in
  the line, which fails if the submodule with the wasm spect testsuite
  is missing. To avoid this, we add the `?submodules=` arg to that call,
  causing nix not to discard of our submodule.

Signed-off-by: wucke13 <wucke13+github@gmail.com>