fix(spec-tests): improve testsuite result comparison
Prior to this change, the testsuite runner executed `assert_return`
directives roughly like this:
1. Convert the expected return values of type `WastRet` to our
   equivalent `Value` types.
2. Execute our interpreter, producing `Value` results.
3. Compare the actual to the expected results.

However, a workaround had to be used for NaN values. The `WastRet` type
allows for two special NaN patterns: `ArithmeticNaN` and `CanonicalNaN`.
This commit changes the testsuite runner, so that `WastRet` types
can be compared directly to the `Value`s produced by our interpreter,
essentially getting rid of the intermediate conversion step. This is the
same approach wasmi uses[1].

[1]: https://github.com/wasmi-labs/wasmi/blob/774995bf37cd7d5a91ecbe1337232af4cf1b4d8f/crates/wast/src/lib.rs#L371

Signed-off-by: Florian Hartung <florian.hartung@dlr.de>