refactor: remove all `CustomResult` type aliases
Type aliases, such as `use X as Y;`, should be used sparingly. They make
code harder to read because they allow modules to use a different names
for the same things. In our case, there is no necessity for them,
especially not for `Result` types, so this commit removes them.

Signed-off-by: Florian Hartung <florian.hartung@dlr.de>