refactor(core): flatten errors for unsupported proposals
Currently, we are only able to detect the use of the proposal for
multiple memories. The other `Proposal` enum variants were unused and
AFAIK there no plan to detect them in a general way as of now. Thus,
flattening the error type hierarchy by directly adding an error for when
multiple memories are used, is sensible.

Signed-off-by: Florian Hartung <florian.hartung@dlr.de>