fix(core): remove `WasmReadable::read_unvalidated`
This removes the second `read_unvalidated` method from the
`WasmReadable` trait and all implementors. The method was originally
introduced with the hope that it would be faster during execution,
because it can directly panic instead of having to bubble up errors.

However, it causes code duplication for all implementors while not
really providing any major benefits. The performance gain could even
be negligible if reader functions are inlined and error paths optimized
away.

Signed-off-by: Florian Hartung <florian.hartung@dlr.de>