fix(core,execution): make `extern_type` methods infallible and panic on incorrect use
In the following the change for `ImportDesc::extern_type` is
described. However, the same reasoning applies to the changes made to
`ExternVal::extern_type`.

The `ImportDesc::extern_type` method returns the `ExternType` of `self`.
In the case of `ImportDesc::Func`, a functype lookup is performed, thus
requiring `ValidationInfo` as an argument. `ImportDesc::extern_type`
is fallible, suggesting to the user that incorrect usage will return
an `Err`.

However, passing an incorrect `ValidationInfo`, i.e. one different from
the `ValidationInfo` used to create `self`, results in an error only
when the accessed function's type index is out of bounds for the passed
`ValidationInfo`. Otherwise the method returns an incorrect `Ok` result.

This method's interface is not intuitive and returned errors can never
be handled as they are a bug in the program. Therefore, this commit
makes the method infallible and panic on invalid accesses instead.

Signed-off-by: Florian Hartung <florian.hartung@dlr.de>