refactor(spec-tests): replace `WastTestReport` with `Result`
`WastTestReport` is an enum that is created per .wast file. It is either
a number of assertion results `AssertReport` or a `ScriptError` if
parsing or some other non-assert directive failed.

This commit removes the `WastTestReport` enum and replaces it with
an equivalent `Result<AssertReport, ScriptError>` type. This new type
allows the use of the ? operator for `ScriptError`s, for which a not so
pretty `try_to!` macro was required before.

Signed-off-by: Florian Hartung <florian.hartung@dlr.de>