refactor(spec-tests): new `AssertOutcome`
`WastSuccess` and `WastError` are used as a `Result<WastSuccess,
WastError>` to store keep track of the outcome of individual assert
directives. Both types store the line number and a context string
as fields. The `WastError` type additionally stores the error
produced, while there is no additional information associated with the
`WastSuccess` type.

Therefore, to reduce duplication, this commit combines both variants
into one new struct `AssertOutcome`. It now stores the line number and
context and an Option for the error in the case where the assert failed.

Signed-off-by: Florian Hartung <florian.hartung@dlr.de>