fix: pull `Option` out of func/extern addrs
The `FuncAddr` and `ExternAddr` used to store a `Option<usize>` each,
which would be the respective addr or `None`, if the reference that
contained these types was null.

However, the spec does not define both addrs as nullable, but instead
defines a null variant directly on the ref type (ref.null) and defines
all addr types as simple primitive integers.

This commit pulls out the `Option`s out of both addr types and adds a
new variant `Null` on the `Ref` value type. This also allows to simplify the
`InteropValue` implementation for function addrs, as they are
now simply `Option<FuncAddr>`s.

In the future, a newtype struct is added for `Option<FuncAddr>` to make
the the API a bit more intuitive.

Signed-off-by: Florian Hartung <florian.hartung@dlr.de>