refactor: move `little_endian` to execution
The `LittleEndianBytes` trait is used exclusively by the execution
module. Moving it from the core to the execution module makes sense and
also allows it to depend on the `Value` type, which is only available in
the execution module. This is necessary for a future rework of Wasm <->
Rust type conversions.

Signed-off-by: Florian Hartung <florian.hartung@dlr.de>