fix(execution): use float newtypes instead of primitives
We have two newtypes `F32` and `F64` for Wasm float values. They add
additional methods not included in Rust's `core` library. To improve
code consistency we should not use Rust's primitive float types. This
also prevents discussions for whether an `f32` or a `F32` would be more
suitable in which scenarios.

This commit replaces primitive floats with our float newtypes in
execution and its unit tests. However, integration tests are excluded as
they test the system from an outside perspective. Thus they are allowed
to use the `InteropValue` traits for conversions.

Signed-off-by: Florian Hartung <florian.hartung@dlr.de>