fix(execution): make Wasm to Rust value conversion fallible
Previously the `From` trait was used to convert Wasm `Value`s to Rust
values. However, this conversion uses `unwrap_validated` to panic when
both the source and target types do not match.

The `core::convert::From` documentation says: "This trait must not fail.
The From trait is intended for perfect conversions. If the conversion
can fail or is not perfect, use TryFrom."

Because our conversion from a `Value` to a Rust value is not perfect,
the fallible `TryFrom` trait should be used, which is what this
commit fixes. Now, whenever a conversion is performed, the error case
has to be dealt with. For us this means that we require an explicit
`unwrap_validated` in the interpreter loop for every conversion.

One could argue, that this increases code bloat drastically, however
I think it is quite nice to have explicit `unwrap_validated`s directly
in the interpreter loop instead of hidden unwraps in some generic
conversion function.

Note that the conversion in the opposite direction, i.e. Rust to Wasm
value, is still using the infallible `From` trait, which is fine.

Signed-off-by: Florian Hartung <florian.hartung@dlr.de>