refactor: move errors to where they are used
The `RuntimeError` and `TrapError` types are used only by the
`execution` module. Therefore it makes sense to also define them there.

Also, the opt-in `Error` and `Result` types must never be used by us, so
it's probably the best if they are defined in the top-level `lib.rs` or
maybe in a top-level `error.rs` module.

Signed-off-by: Florian Hartung <florian.hartung@dlr.de>