fix: trade off panic with wasted slot in slotmap
This change wastes a slotmap slot intentionally when the slot has run
out of possible generations provide for keys, instead of panicking on an
insertion. The generation increments are also delayed to the removal of
slots, this way we can avoid the worst case scenario of linear lookup of
an unoccupied slot that has not run out of fresh generations to provide.

Signed-off-by: Cem Onem <cem.oenem@dlr.de>