wip: dummy sidetable entry for loop instructions
These will store the basic block data for the basic block following
this loop. Note that this instruction is not to be visited again by a
backwards pointing branch instr, so when we do branch off back to the
beginning of the loop the sidetable entry ABOVE the newly updated stp
should be checked.

Signed-off-by: Cem Onem <cem.oenem@dlr.de>