fix(core): remove `WasmReadable` trait
The `WasmReadable` trait was generally a bad design choice. The trait is
never used as a bound for generic types and it just forces implementors
to use the given function signature -- one which does not allow for
context to be passed in.

In order to rework the readers for indices and other types this trait
has to be removed, which is what this commit does.

Signed-off-by: Florian Hartung <florian.hartung@dlr.de>