fix: replace incorrect funcaddrs with externaddrs in table tests
There are multiple tests for the table instructions, which are basically
rewrites of some `*.wast` tests in Rust code.

Back when these were originally written, our interpreter did not
have proper support for externaddrs, which is why funcaddrs were
used instead. However, the usage of invalid/unchecked funcaddrs
(which essentially point to non-exisiting functions) turns out to be
problematic for future implementations. Also, we now have proper support
for externaddrs.

This commit fixes the table tests and replaces incorrect uses of
funcaddrs with externaddrs, just as it was originally done in the
respective `*.wast` tests.

Signed-off-by: Florian Hartung <florian.hartung@dlr.de>