feat: config methods for fuel for each instruction
This commit adds Config::*_flat_cost() methods in which the user can
supply the fuel to be deducted when a particular instruction is fetched.
In order to make sure simple implementations that match the instructions
into constants are folded into a constant themselves in the interpreter
loop the method is called in each match arm in the interpreter fetch
loop instead of once at the beginning.

Signed-off-by: Cem Onem <cem.oenem@dlr.de>